스터디 만들기 수정된 레이아웃
